# Improve DateTime{Offset}.ParseExact{Multiple} performance for RFC1123 ("r")