Significantly improves the performance of `DateTime.ParseExact(..., "r", ...)` by porting and adapting the `Utf8Parser` code from corefx, for a **15-20x** throughput increase (and also saving three string allocations totaling ~80bytes... which more generally we should subsequently look at getting rid of for other formats).  This optimizes for the (default) case of a `DateTimeStyles.None`; specifying any other style falls back to the normal parsing support, as that requires handling things such as arbitrary whitespace anywhere in the string.

Benchmark:
```C#
using System;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Attributes.Jobs;
using BenchmarkDotNet.Running;

[MemoryDiagnoser]
[InProcess]
public class Benchmark
{
    private static void Main() => BenchmarkRunner.Run<Benchmark>();

    [Benchmark] public DateTime DT_ParseExactR() => DateTime.ParseExact("Wed, 15 Aug 1906 07:24:05 GMT", "r", null);
    [Benchmark] public DateTimeOffset DTO_ParseExactR() => DateTimeOffset.ParseExact("Wed, 15 Aug 1906 07:24:05 GMT", "r", null);

    private static readonly string[] s_formats = new string[] { "R" };
    [Benchmark] public DateTime DT_ParseExactMultipleR() => DateTime.ParseExact("Wed, 15 Aug 1906 07:24:05 GMT", s_formats, null);
    [Benchmark] public DateTimeOffset DTO_ParseExactMultipleR() => DateTimeOffset.ParseExact("Wed, 15 Aug 1906 07:24:05 GMT", s_formats, null);
}
```

Before:
```
                  Method |     Mean |     Error |    StdDev |  Gen 0 | Allocated |
------------------------ |---------:|----------:|----------:|-------:|----------:|
          DT_ParseExactR | 1.922 us | 0.0377 us | 0.0888 us | 0.0172 |      80 B |
         DTO_ParseExactR | 1.948 us | 0.0374 us | 0.0511 us | 0.0153 |      80 B |
  DT_ParseExactMultipleR | 1.922 us | 0.0265 us | 0.0221 us | 0.0172 |      80 B |
 DTO_ParseExactMultipleR | 1.934 us | 0.0349 us | 0.0292 us | 0.0153 |      80 B |
```

After:
```
                  Method |      Mean |     Error |    StdDev | Allocated |
------------------------ |----------:|----------:|----------:|----------:|
          DT_ParseExactR |  98.38 ns | 0.9119 ns | 0.7119 ns |       0 B |
         DTO_ParseExactR | 123.87 ns | 0.4554 ns | 0.3012 ns |       0 B |
  DT_ParseExactMultipleR | 112.05 ns | 2.3175 ns | 4.6284 ns |       0 B |
 DTO_ParseExactMultipleR | 120.27 ns | 1.3823 ns | 1.2930 ns |       0 B |
```

cc: @ahsonkhan, @jkotas, @pjanotti, @joperezr 
Contributes to https://github.com/dotnet/corefx/issues/30612