@stephentoub 2018-07-03
@tarekgh or @jkotas, do you have any familiarity with the issue cited at:
https://github.com/dotnet/coreclr/blob/dc8ee8c84cf51a474fbfc164f7efe825a7a7aece/src/System.Private.CoreLib/shared/System/Globalization/DateTimeParse.cs#L4220-L4230
and whether that's something I need to care about here?