@tarekgh 2018-07-03
@stephentoub 

> do you have any familiarity with the issue cited at:

This is to support the date format with a quote. i.e. in the date format, you'll have 'z' or 'GMT'. the quotes are not part of the real ISO format but it looks the framework at some time supported it by mistake. usually, when having quotes in the date pattern, the characters inside the quotes are treated literals (i.e. it show up in the formatted strings as it is) but when having 'z' or 'GMT' we treat it is as the date is in UTC time zone.