@stephentoub 2018-07-04
> This is to support the date format with a quote

@tarekgh, can you share an example?  I've not been able to come up with a variant that has quotes around GMT and that successfully parses with DateTime.ParseExact(..., "r", ...)