@stephentoub 2018-07-04
Oh, wait, it's to support the date format, not the actual string.  So I don't need to worry about that here, then.  Thanks.