@tarekgh 2018-07-04
@stephentoub please ensure you parse the following correctly: 

            string s = "Mon, 01 Jan 0001 00:00:00 'GMT'";
            dt = DateTime.ParseExact(formatted, "r", CultureInfo.CurrentCulture);

This is the special case we support. note that the quotes around GMT in the formatted string.