@stephentoub 2018-07-04
> please ensure you parse the following correctly:

@tarekgh, this is what I was asking about earlier.  That example fails for me, e.g. this program:
```C#
using System;
using System.Globalization;

class Program
{
    static void Main() => DateTime.ParseExact("Mon, 01 Jan 0001 00:00:00 'GMT'", "r", CultureInfo.CurrentCulture);
}
```
throws on .NET 4.7.2:
```
Unhandled Exception: System.FormatException: String was not recognized as a valid DateTime.
   at System.DateTimeParse.ParseExact(String s, String format, DateTimeFormatInfo dtfi, DateTimeStyles style)
   at System.DateTime.ParseExact(String s, String format, IFormatProvider provider)
```
Can you please clarify / share an example that parses successfully with format "r" and with quotes around GMT?

> this looks wrong as GMT can specified with quotes so it can be 31

Same question.  I've not been able to come up with an example where quotes are parsed successfully.